services/gateway/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
httpx>=0.27,<1.0
PyJWT[crypto]>=2.8,<3.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
